requirements.txt
acres==0.5.0
altgraph==0.17.4
certifi==2025.10.5
charset-normalizer==3.4.4
ci-info==0.3.0
click==8.3.0
configobj==5.0.9
configparser==7.2.0
etelemetry==0.3.1
filelock==3.20.0
fitz==0.0.1.dev2
httplib2==0.31.0
idna==3.11
looseversion==1.3.0
lxml==6.0.2
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
prov==2.1.1
puremagic==1.30
pydot==4.0.1
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
PyMuPDF==1.26.5
pyparsing==3.2.5
pyperclip==1.11.0
python-dateutil==2.9.0.post0
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.3
rdflib==7.2.1
requests==2.32.5
scipy==1.16.2
setuptools==80.9.0
simplejson==3.20.2
six==1.17.0
traits==7.0.2
tzdata==2025.2
urllib3==2.5.0
